+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/RagManager.java
@@ -76,7 +76,7 @@ public class RagManager
         new HashMap<Object,List<Transaction>>();
 
     private final ArrayMap<Transaction,Object> waitingTxMap =
-        new ArrayMap<Transaction,Object>( 5, false, true );
+        new ArrayMap<Transaction,Object>( (byte)5, false, true );
 
     private final TransactionManager tm;
